refactor Import screen into components
