+++ src/frontend/screens/Onboarding/walletSelection/screens/import/components/BrowserPackageManagerImportOptions/index.module.scss
@@ -1,0 +1,69 @@
+.packageOptionsContainer {
+  background-color: var(--color-neutral-700);
+  padding: var(--space-lg);
+  border-radius: var(--space-sm);
+}
+
+.packageManagerSubTitle {
+  margin-bottom: var(--space-md);
+}
+
+.importBrowserProfileOption {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  background-color: var(--color-neutral-800);
+  border-radius: var(--space-sm);
+  padding: var(--space-md);
+  gap: var(--space-sm);
+  border: 1px solid var(--color-stroke-01);
+  margin: var(--space-2xs) 0;
+
+  &:hover {
+    border: 1px solid #626161;
+    opacity: 100%;
+    background-color: var(--color-neutral-600);
+  }
+}
+
+.profileImage {
+  border-radius: var(--space-sm);
+  overflow: hidden;
+  width: 48px;
+  height: 48px;
+  img {
+    width: 100%;
+  }
+}
+
+.defaultProfileImage {
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  > svg {
+    height: 20px;
+  }
+}
+
+@mixin hideLongText {
+  text-overflow: ellipsis;
+  max-width: 200px;
+  overflow-x: hidden;
+}
+
+.profileName {
+  text-align: left;
+  @include hideLongText;
+}
+
+.displayName {
+  color: var(--color-neutral-400);
+  @include hideLongText;
+}
+
+.profileRightArrow {
+  justify-self: flex-end;
+  margin: 0 0 0 auto;
+}
