+++ src/frontend/screens/Onboarding/walletSelection/screens/import/components/ImportAndCreateOptions/index.module.scss
@@ -1,0 +1,12 @@
+@use '../../mixins';
+
+.importOptionsContainer {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  grid-auto-rows: 112px;
+  gap: 16px;
+  margin: var(--space-sm) auto calc(2 * var(--space-sm));
+  max-height: 300px;
+  overflow-y: scroll;
+  @include mixins.showScrollbar;
+}
