+++ src/frontend/screens/Onboarding/walletSelection/screens/import/index.module.scss
@@ -1,22 +1,3 @@
-@mixin showScrollbar {
-  &::-webkit-scrollbar {
-    width: 20px;
-    height: 20px;
-    background: transparent;
-  }
-}
-
-.importOptionsContainer {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-auto-rows: 112px;
-  gap: 16px;
-  margin: var(--space-sm) auto calc(2 * var(--space-sm));
-  max-height: 300px;
-  overflow-y: scroll;
-  @include showScrollbar;
-}
-
 .title {
   align-self: center;
   margin-bottom: var(--space-sm);
@@ -31,136 +12,4 @@
 
 .errorMessage {
   margin: 0px auto;
-}
-
-.packageOptionsContainer {
-  background-color: var(--color-neutral-700);
-  padding: var(--space-lg);
-  border-radius: var(--space-sm);
-}
-
-.packageManagerSubTitle {
-  margin-bottom: var(--space-md);
-}
-
-.importBrowserProfileOption {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  background-color: var(--color-neutral-800);
-  border-radius: var(--space-sm);
-  padding: var(--space-md);
-  gap: var(--space-sm);
-  border: 1px solid var(--color-stroke-01);
-  margin: var(--space-2xs) 0;
-
-  &:hover {
-    border: 1px solid #626161;
-    opacity: 100%;
-    background-color: var(--color-neutral-600);
-  }
-}
-
-@mixin hideLongText {
-  text-overflow: ellipsis;
-  max-width: 200px;
-  overflow-x: hidden;
-}
-
-.displayName {
-  color: var(--color-neutral-400);
-  @include hideLongText;
-}
-
-.profileRightArrow {
-  justify-self: flex-end;
-  margin: 0 0 0 auto;
-}
-
-.profileImage {
-  border-radius: var(--space-sm);
-  overflow: hidden;
-  width: 48px;
-  height: 48px;
-  img {
-    width: 100%;
-  }
-}
-
-.goBackProfileButton {
-  max-width: 120px;
-  color: var(--color-neutral-300);
-}
-
-.manualFolderPickerContainer {
-  margin-top: var(--space-lg);
-}
-
-.folderPickerButton {
-  width: 100%;
-  text-overflow: ellipsis;
-  color: var(--color-neutral-300);
-  text-align: left;
-}
-
-.importSubtitle {
-  margin-bottom: var(--space-2xs);
-  color: var(--color-neutral-400);
-}
-
-.cantFindSubtext {
-  margin-top: var(--space-md);
-  display: inline-block;
-  > div {
-    display: inline-block;
-  }
-}
-
-@mixin actionContainer {
-  margin: var(--space-md) auto;
-}
-
-.importActionContainer {
-  @include actionContainer;
-}
-
-.manualImportActionContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  @include actionContainer;
-  width: 100%;
-}
-
-.manualHr {
-  border: none;
-  height: 1px;
-  background-color: var(--color-stroke-01);
-  margin: var(--space-lg) 0 0 0;
-}
-
-.profileOptionsContainer {
-  overflow-x: hidden;
-  overflow-y: scroll;
-  max-height: 240px;
-  @include showScrollbar;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-2xs);
-}
-
-.profileName {
-  text-align: left;
-  @include hideLongText;
-}
-
-.defaultProfileImage {
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  > svg {
-    height: 20px;
-  }
 }
